refactor: change type of SampleSet to uint8_t

+++ include/osu!parser/Parser/Structures/Beatmap/Objects/TimingPoint.hpp
@@ -5,7 +5,7 @@
 
 namespace OsuParser::Beatmap::Objects::TimingPoint
 {
-    enum class SampleSet : std::int32_t
+    enum class SampleSet : std::uint8_t
     {
         NO_CUSTOM = 0,
         NORMAL = 1,
